LayoutTests/storage/domstorage/localstorage/delete-defineproperty-removal.html

<pre id='console'></pre>
<script>
function runTest() {
    function log(msg)
    {
        document.getElementById('console').append(msg + '\n');
    }

    function logAndEval(str)
    {
        log(str);
        eval(str);
    }

    if (window.testRunner)
       testRunner.dumpAsText();

    if (!window.localStorage) {
        log("window.localStorage DOES NOT exist");
        return;
    }

    logAndEval("localStorage.clear()");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")\n");
    
    logAndEval("Object.defineProperty(localStorage, '1', {value: 'present'})");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")");
    // FIXME: This is incorrect. It should return 'present'. See https://webkit.org/b/172687.
    log("localStorage.getItem('1') is " + localStorage.getItem('1') + "\n");
    
    logAndEval("delete localStorage[1]");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")\n");
    
    logAndEval("Object.defineProperty(localStorage, '2', {value: 'present',  configurable: true})");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")");
    // FIXME: This is incorrect. It should return 'present'. See https://webkit.org/b/172687.
    log("localStorage.getItem('2') is " + localStorage.getItem('2') + "\n");
    
    logAndEval("delete localStorage[2]");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")\n");

    logAndEval("Object.defineProperty(localStorage, 'Test', {value: 'present',  configurable: true})");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")");
    // FIXME: This is incorrect. It should return 'present'. See https://webkit.org/b/172687.
    log("localStorage.getItem('Test') is " + localStorage.getItem('Test') + "\n");
  
    logAndEval("delete localStorage['Test']");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")\n");

    logAndEval("Object.defineProperty(localStorage, 'Test1', {value: 'present'})");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")");
    // FIXME: This is incorrect. It should return 'present'. See https://webkit.org/b/172687.
    log("localStorage.getItem('Test1') is " + localStorage.getItem('Test1') + "\n");
  
    logAndEval("delete localStorage['Test1']");
    log("Object.getOwnPropertyNames(localStorage) is (" + Object.getOwnPropertyNames(localStorage) + ")\n");
}
</script>
<body onload="runTest();">
</body>
